requirements.txt
langchain
langchain-openai
langchain-google-genai
langgraph
pylint
pytest
python-dotenv
pandas
colorama